modify model relationship
