+++ src/migrate_db.py
@@ -1,5 +1,5 @@
 from sqlalchemy import create_engine
-from src.models.task import Base
+from src.models.model import Base
 
 DB_URL = "mysql+pymysql://root@db:3306/todo?charset=utf8mb4"
 engine = create_engine(DB_URL, echo=True)
