--- src/models/task.py
@@ -1,14 +1,0 @@
-from sqlalchemy import Boolean, Column, Integer, String
-from sqlalchemy.orm import relationship
-from sqlalchemy.sql.schema import ForeignKey
-from src.db import Base
-
-class Task(Base):
-    __tablename__ = "tasks"
-
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    title = Column(String(1024))
-    done = Column(Boolean)
-    owner_id = Column(Integer, ForeignKey("users.id"))
-
-    owner = relationship("User", back_populates="tasks")
